added speed slider, fixed sliders, added button

+++ FlockingSimulation/src/tools/Utils.java
@@ -2,9 +2,9 @@ package tools;
 
 public class Utils {
 	
-	public static void pause(int time) {
+	public static void pause(double time) {
 		try {
-			Thread.sleep(time);
+			Thread.sleep((long)time);
 		} catch (InterruptedException e) {
 			// We are happy with interruptions, so do not report exception
 		}
